backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
langchain==0.1.0
langchain-google-genai==0.0.14
python-dotenv==1.0.0
langchain-text-splitters==1.1.0
